Module rust_wasm_dodrio_templating::html_template_mod

html_template_mod
Html templating for dodrio, generic code for a standalone library. The implementation is in another file where RootRenderingComponents implement the trait HtmlTemplating

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

Traits

HtmlTemplating
the RootRenderingComponent struct must implement this trait it must have the fields for local_route and html_template fields

Functions

decode_5_xml_control_characters
decode 5 xml control characters : " ' & < >
https://www.liquid-technologies.com/XML/EscapingData.aspx I will ignore all html entities, to keep things simple, because all others characters can be written as utf-8 characters. https://www.tutorialspoint.com/html5/html5_entities.htm

empty_div
get en empty div node