Trait rust_wasm_dodrio_templating::html_template_mod::HtmlTemplating

pub trait HtmlTemplating {
    fn replace_with_string(&self, fn_name: &str) -> String;

    fn retain_next_node_or_attribute<'a>(&self, fn_name: &str) -> bool;

    fn replace_with_nodes<'a>(
        &self, 
        cx: &mut RenderContext<'a>, 
        fn_name: &str
    ) -> Vec<Node<'a>>;

    fn set_event_listener(
        &self, 
        fn_name: String
    ) -> Box<dyn Fn(&mut dyn RootRender, VdomWeak, Event) + 'static>;

    fn render_template<'a>(
        &self, 
        cx: &mut RenderContext<'a>, 
        html_template: &str, 
        html_or_svg_parent: HtmlOrSvg
    ) -> Result<Node<'a>, String> { ... }

    fn fill_element_builder<'a>(
        &self, 
        reader_for_microxml: &mut ReaderForMicroXml<'_>, 
        element: ElementBuilder<'a, Vec<'a, Listener<'a>>, Vec<'a, Attribute<'a>>, Vec<'a, Node<'a>>>, 
        cx: &mut RenderContext<'a>, 
        html_or_svg_parent: HtmlOrSvg, 
        dom_path: &mut Vec<String>
    ) -> Result<ElementBuilder<'a, Vec<'a, Listener<'a>>, Vec<'a, Attribute<'a>>, Vec<'a, Node<'a>>>, String> { ... }
}

the RootRenderingComponent struct must implement this trait it must have the fields for local_route and html_template fields

Required methods

fn replace_with_string(&self, fn_name: &str) -> String

fn retain_next_node_or_attribute<'a>(&self, fn_name: &str) -> bool

fn replace_with_nodes<'a>(
&self,
cx: &mut RenderContext<'a>,
fn_name: &str
) -> Vec<Node<'a>>

fn set_event_listener(
&self,
fn_name: String
) -> Box<dyn Fn(&mut dyn RootRender, VdomWeak, Event) + 'static>

Loading content...
Provided methods

fn render_template<'a>(
&self,
cx: &mut RenderContext<'a>,
html_template: &str,
html_or_svg_parent: HtmlOrSvg
) -> Result<Node<'a>, String>

get root element Node.
I wanted to use dodrio::Node, but it has only private methods.
I must use dodrio element_builder.

fn fill_element_builder<'a>(
&self,
reader_for_microxml: &mut ReaderForMicroXml<'_>,
element: ElementBuilder<'a, Vec<'a, Listener<'a>>, Vec<'a, Attribute<'a>>, Vec<'a, Node<'a>>>,
cx: &mut RenderContext<'a>,
html_or_svg_parent: HtmlOrSvg,
dom_path: &mut Vec<String>
) -> Result<ElementBuilder<'a, Vec<'a, Listener<'a>>, Vec<'a, Attribute<'a>>, Vec<'a, Node<'a>>>, String>

Recursive function to fill the Element with attributes and sub-nodes(Element, Text, Comment).
Moves & Returns ElementBuilder or error.
I must move ElementBuilder because its methods are all move.
It makes the code less readable. It is only good for chaining and type changing.

Loading content...